druga izmjena

+++ student/app/Http/Controllers/PollController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\View;
+use Illuminate\Support\ServiceProvider;
 use Illuminate\Http\Request;
 use RealRashid\SweetAlert\Facades\Alert;
 use App\Poll;
@@ -14,9 +16,12 @@ use DB;
 
 class PollController extends Controller
 {
+    
+    
     //Retriving all poll from database
     public function allPoll()
     {
+        
         $polls = Poll::all();
         
         return view('poll.list_all_poll',compact('polls'));
@@ -38,34 +43,60 @@ class PollController extends Controller
      */
     public function create(Request $request)
     {
-        $validatePoll = $this->validate($request, [
+        $messages = [
+            'question.required' => 'Question cannot be empty',
+            'option.required' => 'Option cannot be empty',
+            'start_at.required' => 'Start date cannot be empty',
+            'end_at.required' => 'End date cannot be empty'
+        ];
+        
+        $this->validate($request, [
             'question' => 'required',
             'option' => 'required|array',
             'start_at' => 'required',
             'end_at' => 'required'
         ]);
+        $validatePoll = $request->all();
+
+        $start_at = request('start_at');
+        $end_at = request('end_at');
         
-        if($validatePoll){
-            $poll = new Poll();
-            $poll->user_id = request()->user()->id;
-            $poll->question = request()->question;
-            $poll->start_at = request()->start_at;
-            $poll->end_at = request()->end_at;
-            $poll->isActive = 2;
-            $options = request()->option;
-         
-            if($poll->save()){
-                foreach ($options as $option){
-                    $poll->option_polls()->createMany([
-                        ['option' => $option], 
-                    ]);
+        $countPoll = Poll::where(function ($query) use ($start_at, $end_at) {
+            $query->where(function ($query) use ($start_at, $end_at) {
+               $query->where('start_at', '>=', $start_at)
+                       ->where('end_at', '<', $start_at);
+               })
+               ->orWhere(function ($query) use ($start_at, $end_at) {
+                   $query->where('start_at', '<', $end_at)
+                           ->where('end_at', '>=', $end_at);
+               });
+           })->count();
+
+           if($countPoll > 0) {
+             toast("Please use another date !",'warning');
+           } else {
+            if($validatePoll){
+                $poll = new Poll();
+                $poll->user_id = request()->user()->id;
+                $poll->question = request()->question;
+                $poll->start_at = request()->start_at;
+                $poll->end_at = request()->end_at;
+                $poll->isActive = 2;
+                $options = request()->option;
+             
+                if($poll->save()){
+                    foreach ($options as $option){
+                        $poll->option_polls()->createMany([
+                            ['option' => $option], 
+                        ]);
+                    }
+                } else {
+                    toast("Something wrong !",'warning');
                 }
-                Alert::success("Success ! Poll created !");
-            } else {
-                Alert::warning("Oops ! Something wrong !");
-            }
-            Alert::success("Success ! Poll Created!");
-        }  
+                toast("Poll created !",'success');
+            }  
+           }
+        
         return redirect('/poll');
     }
     
@@ -79,8 +110,9 @@ class PollController extends Controller
     {
         $poll = Poll::find($id);
         $options = $poll->option_polls;
+        $sum  = $poll->option_polls()->sum('vote');
         
-        return view('poll.edit_poll', compact('poll', 'options'));
+        return view('poll.edit_poll', compact('poll', 'options', 'sum'));
     }
 
     /**
@@ -91,7 +123,7 @@ class PollController extends Controller
     {
         $poll = Poll::find($id);
         
-        if($poll->isActive == 1 || $poll->isActive == 2 ) {
+        if($poll->isActive == 1 ) {
             $poll->end_at = request('end_at');
         } else {
             $poll->user_id = request()->user()->id;
@@ -99,25 +131,54 @@ class PollController extends Controller
             $poll->start_at = request()->start_at;
             $poll->end_at = request()->end_at;
             $options = request()->option;
-
+            
             if($poll->save()){
-                $poll->option_polls()->associate(request()->option);
+                $poll->option_polls()->delete();
 
-                Alert::success('Success', 'Poll updated !');
+                foreach ($options as $option){
+                    $poll->option_polls()->createMany([
+                        ['option' => $option], 
+                    ]);
+                }
+
+                toast('Poll updated !','success');
             } else {
-                Alert::warning('Ooops', 'Something wrong !');
+                toast('Something wrong !','error');
             }
         }
-        return view('/poll/' .$poll->id. '/edit');
+        return redirect('/poll/' .$poll->id. '/edit');
     }
 
-   //Close active or waiting poll (isActive = 0)
-    public function closePoll($id){}
-
-    public function checkWaiting()
-    {
-        $waitingPoll = Poll::where('isActive', 2)->firstOrFail()->update(['isActive' => 1]);
+   /**
+    * Close active or waiting poll (isActive = 0)
+   */
+    public function closePoll($id){
+        $poll = Poll::find($id);
         
+        $poll->isActive = 0;
+       
+       if($poll->save()) {
+        toast('Poll is closed!','success');
+       } else {
+        toast('Something wrong !','warning');
+       }
+
+        return redirect('/list_poll');
+        
+    }
+
+    public function activePoll()
+    {   
+        $poll = Poll::where('isActive',2)
+                    ->orWhere(function($query) {
+                        $query->whereDate('start_at',Carbon::now());
+                    })->first(); 
+        
+        if($poll != null) {
+            $poll->isActive = 1;
+            $poll->save();
+        } 
+        //return compact('poll');
     }
 
     public function checkActive()
@@ -126,29 +187,32 @@ class PollController extends Controller
         $options = $activePoll->option_polls()->sum('vote');
         
         return ([
+            'activePoll' => $activePoll,
             'options' => $options, 
         ]);
-
     }
+
     /**
      * Delete poll older than 15 days
      */
      public function deleteOldPoll()
      {
         $polls = Poll::where('isActive', '0')->get();
-        
+          if(count($polls) > 0) {
             foreach($polls as $poll) {
-                $endDate = $poll->created_at;
-               
-                if($endDate->subDays(10)->diffInDays() > 5 ){
-                    
-                    $poll->delete();
-                    $poll->option_polls()->delete();
-                   //testna izmjena
-                }
-            }
+              $endDate = $poll->created_at;
             
-        
+              if($endDate->subDays(10)->diffInDays() > 5 ){
+                  
+                $poll->delete();
+                $poll->option_polls()->delete();
+                
+               toast('All Events are deleted!','success');
+              }
+            }
+          } else {
+            toast('Nothing to delete!','warning');
+          }
         return redirect('list_poll');
      }
 
@@ -165,12 +229,36 @@ class PollController extends Controller
             $votes = request('option');
             $voteIncrement = DB::table('option_polls')->where('id', $votes)->increment('vote');
 
-            Alert::success("Thank you for vote !");
+            toast("Thank you for vote !",'success');
             $vote = ['voted' =>  1, 'user_id' => auth()->user()->id ];
             
             $isVoted = json_encode($vote);
             
             return redirect()->back()->cookie('voted',$isVoted,30);
         }
+    }
+
+    public function getResult($id) {
+        $poll = Poll::find($id);
+
+        $options = $poll->option_polls;
+
+        $sum =  $poll->option_polls()->sum('vote');
+
+        return view('poll.poll-result', compact('poll','options','sum'));
+    }
+
+    public function delete($id) 
+    {
+        $poll = Poll::find($id);
+        if($poll->isActive == 1) {
+            toast('Poll is active! Cant delete !','warning');
+        } else {
+            $poll->delete();
+            $poll->option_polls()->delete();
+
+            Alert::success('Success', 'Deleted !');
+        }
+        return redirect('/list_poll');
     }
 }
